requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
scikit-learn==1.4.2
transformers==4.41.2
sentence-transformers==2.7.0
pandas==2.2.2
numpy==1.26.4
mlflow==2.14.1
SQLAlchemy==2.0.30
psycopg[binary]==3.1.19
python-dotenv==1.0.1
orjson==3.10.3
joblib==1.4.2
datasets==2.19.1
accelerate>=0.21.0
torch>=2.2.0
packaging>=24.0
typer==0.12.3
boto3==1.34.145
moto[s3]==5.1.13